third_party/blink/web_tests/svg/dynamic-updates/SVGImageElement-dom-preserveAspectRatio-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'preserveAspectRatio' attribute of the SVGImageElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS imageElement.getAttribute('preserveAspectRatio') is "xMaxYMax meet"
PASS imageElement.getAttribute('preserveAspectRatio') is "none"
PASS successfullyParsed is true

TEST COMPLETE
